+++ cppwinrt/file_writers.h
@@ -9,7 +9,6 @@ namespace cppwinrt
         write_open_file_guard(w, "BASE");
 
         w.write(strings::base_includes);
-        w.write(strings::base_coroutine);
         w.write(strings::base_macros);
         w.write(strings::base_types);
         w.write(strings::base_extern);
